initial test folder refactor
