+++ contracts/airdrops/TokenDrop.sol
@@ -96,19 +96,19 @@ contract TokenDrop is ERC20, AirdropInfo, AirdropMerkleProof, Ownable {
         _mint(msg.sender, tokensPerClaim * tokenIds.length);
     }
 
-    //@notice Get the type of airdrop, it's either ERC20, ERC721, ERC1155
+    ///@notice Get the type of airdrop, it's either ERC20, ERC721, ERC1155
     function getAirdropType() external pure returns (string memory) {
         return "ERC20";
     }
 
-    //@notice Checks if the user is eligible for this airdrop
+    ///@notice Checks if the user is eligible for this airdrop
     function isEligibleForReward(uint256 tokenId) external view returns (bool) {
         if (hasClaimed[tokenId]) revert AlreadyRedeemed();
         if (rewardedNft.ownerOf(tokenId) != msg.sender) revert NotOwner();
         return true;
     }
 
-    //@notice Returns the amount(number) of airdrop tokens to claim
+    ///@notice Returns the amount(number) of airdrop tokens to claim
     //@param tokenId is the rewarded NFT collections token ID
     function getAirdropAmount() external view returns (uint256) {
         return rewardedNft.balanceOf(msg.sender) * tokensPerClaim;
